separate on classes an functionCreate
